+++ packages/react-storage/src/components/StorageBrowser/components/composables/Navigation.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import { BreadcrumbNavigation } from '../base';
+
+interface NavigationItem {
+  isCurrent?: boolean;
+  name?: string;
+  onNavigate?: () => void;
+}
+
+export interface NavigationProps {
+  items: NavigationItem[];
+}
+
+export const Navigation = ({
+  items,
+}: NavigationProps): React.JSX.Element | null => {
+  if (!items.length) {
+    return null;
+  }
+
+  return <BreadcrumbNavigation breadcrumbs={items} />;
+};