+++ packages/react-storage/src/components/StorageBrowser/providers/store/index.ts
@@ -1,3 +1,0 @@
-export { FileItems } from './files';
-export { StoreProvider, StoreProviderProps } from './StoreProvider';
-export { useStore, UseStoreState } from './useStore';