+++ packages/react-storage/src/components/StorageBrowser/views/hooks/useResolveTableData/useResolveTableData.ts
@@ -1,14 +1,10 @@
 import React from 'react';
 
-import { DataTableProps } from '../../../composables/DataTable';
+import { DataTableProps } from '../../../components';
 
 import { DataTableResolvers } from './types';
 
-export default function useResolveDataTable<
-  K extends string,
-  TItem = {},
-  TProps = {},
->(
+export default function useResolveTableData<K extends string, TItem, TProps>(
   keys: readonly K[] | K[],
   { getCell, getHeader, getRowKey }: DataTableResolvers<K, TProps, TItem>,
   { items, props }: { items?: TItem[]; props: TProps }